requirements.txt
appdirs==1.4.3
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
boto3==1.11.9
botocore==1.14.17
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
coverage==5.1
cryptography==2.9
distlib==0.3.0
docutils==0.15.2
filelock==3.0.12
google-auth==1.14.0
gspread==3.4.2
identify==1.4.14
idna==2.9
jmespath==0.9.5
lxml==4.4.2
more-itertools==8.2.0
nodeenv==1.3.5
packaging==20.3
paramiko==2.7.1
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.2.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.3.0
pynamodb==4.3.1
pyparsing==2.4.7
pytest==5.4.1
pytest-parallel==0.0.10
pytest-sftpserver==1.3.0
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
sentry-sdk==0.14.3
six==1.14.0
tblib==1.6.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.18
wcwidth==0.1.9
